Do not rely on serialization of parsed path ops for partial auto-completion
